+++ Source/ACE.Server/WorldObjects/Corpse.cs
@@ -1,18 +1,21 @@
 using ACE.Database.Models.Shard;
 using ACE.Database.Models.World;
-using ACE.DatLoader;
-using ACE.DatLoader.Entity;
-using ACE.DatLoader.FileTypes;
 using ACE.Entity;
 using ACE.Entity.Enum;
 using ACE.Server.Network.Motion;
-using System.Linq;
+using ACE.Server.Entity.Actions;
+using ACE.Server.Managers;
 
 namespace ACE.Server.WorldObjects
 {
     public class Corpse : Container
     {
         private static readonly UniversalMotion dead = new UniversalMotion(MotionStance.Standing, new MotionItem(MotionCommand.Dead));
+
+        /// <summary>
+        /// Timer's value is in heartbeats, which is every 5 seconds. This value equates for five minutes, including the first heartbeat queued at WO creation.
+        /// </summary>
+        int rotTimer = 55;
 
         /// <summary>
         /// A new biota be created taking all of its values from weenie.
@@ -38,6 +41,24 @@ namespace ACE.Server.WorldObjects
 
             ContainerCapacity = 10;
             ItemCapacity = 120;
+        }
+
+        /// <summary>
+        /// Put in place to clean corpses, until more global cleanup management can be implemented
+        /// </summary>
+        public override void HeartBeat()
+        {
+            --rotTimer;
+
+            if (rotTimer <= 0)
+            {
+                ActionChain selfDestructChain = new ActionChain();
+                selfDestructChain.AddAction(this, () => LandblockManager.RemoveObject(this));
+                selfDestructChain.EnqueueChain();
+                return;
+            }
+
+            QueueNextHeartBeat();
         }
 
         public override ACE.Entity.ObjDesc CalculateObjDesc()
