+++ Source/ACE.Server/WorldObjects/Player_Magic.cs
@@ -2,6 +2,7 @@ using ACE.Entity;
 using ACE.Entity.Enum;
 using ACE.Server.Entity.Actions;
 using ACE.Server.Network.Motion;
+using ACE.Server.Network.GameMessages.Messages;
 
 namespace ACE.Server.WorldObjects
 {
@@ -14,14 +15,34 @@ namespace ACE.Server.WorldObjects
         {
             WorldObject target = CurrentLandblock.GetObject(guidTarget) as WorldObject;
 
-            var turnToMotion = new UniversalMotion(MotionStance.Spellcasting, Location, guidTarget);
-            turnToMotion.MovementTypes = MovementTypes.TurnToObject;
+            if (Guid == guidTarget)
+            {
+                CreatePlayerSpell(guidTarget, spellId);
+            }
+            else
+            {
+                // get angle to target
+                var angle = GetAngle(target);
 
-            ActionChain turnToTimer = new ActionChain();
-            turnToTimer.AddAction(this, () => CurrentLandblock.EnqueueBroadcastMotion(this, turnToMotion));
-            turnToTimer.AddDelaySeconds(1);
-            turnToTimer.AddAction(this, () => CreatePlayerSpell(guidTarget, spellId));
-            turnToTimer.EnqueueChain();
+                float delay = 0;
+                if (angle < 10.0f)
+                {
+                    CreatePlayerSpell(guidTarget, spellId);
+                    return;
+                }
+
+                if (angle < 120.0f) delay = 1.0f;
+                else delay = 1.5f;
+
+                var turnToMotion = new UniversalMotion(MotionStance.Spellcasting, Location, guidTarget);
+                turnToMotion.MovementTypes = MovementTypes.TurnToObject;
+
+                ActionChain turnToTimer = new ActionChain();
+                turnToTimer.AddAction(this, () => CurrentLandblock.EnqueueBroadcastMotion(this, turnToMotion));
+                turnToTimer.AddDelaySeconds(delay);
+                turnToTimer.AddAction(this, () => CreatePlayerSpell(guidTarget, spellId));
+                turnToTimer.EnqueueChain();
+            }
         }
 
         /// <summary>
