+++ Source/ACE.Server/WorldObjects/Player_Melee.cs
@@ -70,51 +70,6 @@ namespace ACE.Server.WorldObjects
             MeleeTarget = null;
         }
 
-        public float GetDistance(WorldObject target)
-        {
-            return Location.DistanceTo(target.Location);
-        }
-
-        public float GetAngle(WorldObject target)
-        {
-            var currentDir = Location.GetCurrentDir();
-            var targetDir = GetDirection(target);
-
-            // get the 2D angle between these vectors
-            return GetAngle(currentDir, targetDir);
-        }
-
-        /// <summary>
-        /// Returns the 2D angle between 2 vectors
-        /// </summary>
-        public float GetAngle(Vector3 a, Vector3 b)
-        {
-            var cosTheta = a.Dot2D(b);
-            var rads = Math.Acos(cosTheta);
-            var angle = rads * (180.0f / Math.PI);
-            return (float)angle;
-        }
-
-        public Vector3 GetDirection(WorldObject target)
-        {
-            // draw a line from current location
-            // to target location
-            var offset = target.Location.Pos - Location.Pos;
-            offset = offset.Normalize();
-
-            return offset;
-        }
-
-        public void Rotate()
-        {
-
-        }
-
-        public void MoveTo()
-        {
-
-        }
-
         public void Attack(WorldObject target)
         {
             if (MeleeTarget == null)
